Extract info request to snow route
