+++ ui/app/controllers/snow/apps/show.js
@@ -1,7 +1,10 @@
 import Ember from 'ember';
 
 export default Ember.ObjectController.extend({
+  needs: ["snow"],
+  base_url: Ember.computed.alias("controllers.snow.model.base_url"),
+
   hasRequests: function() {
-    return this.get("app.requests.today") && this.get("app.requests.yesterday");
+    return this.get("requests.today") && this.get("requests.yesterday");
   }.property()
 });
