+++ ui/app/routes/snow.js
@@ -1,4 +1,13 @@
 import Ember from 'ember';
 import AuthenticatedRouteMixin from 'simple-auth/mixins/authenticated-route-mixin';
 
-export default Ember.Route.extend(AuthenticatedRouteMixin, {});
+export default Ember.Route.extend(AuthenticatedRouteMixin, {
+  model: function() {
+    return Ember.$.get(this._infoUrl());
+  },
+
+  _infoUrl: function() {
+    var adapter = this.container.lookup('adapter:application');
+    return adapter.buildURL("") + "/info";
+  }
+});
