+++ ui/app/templates/snow/apps/_info.hbs
@@ -6,11 +6,11 @@
     <tr>
       <td>
         <b>Today</b><br>
-        {{#if app.requests.today}}{{app.requests.today.count}}{{else}}-{{/if}}
+        {{#if requests.today}}{{requests.today.count}}{{else}}-{{/if}}
       </td>
       <td>
         <b>Yesterday</b><br>
-        {{#if app.requests.yesterday}}{{app.requests.yesterday.count}}{{else}}-{{/if}}
+        {{#if requests.yesterday}}{{requests.yesterday.count}}{{else}}-{{/if}}
       </td>
     </tr>
   </table>
