+++ ui/app/templates/snow/apps/_install.hbs
@@ -7,8 +7,8 @@ gem 'snowagent'
 <b>Heroku:</b>
 <pre>
 heroku config:set \
-  SNOWMANIO_SERVER={{info.base_url}} \
-  SNOWMANIO_SECRET_TOKEN={{app.token}}
+  SNOWMANIO_SERVER={{base_url}} \
+  SNOWMANIO_SECRET_TOKEN={{token}}
 </pre>
 
 <b>Standalone:</b>
@@ -18,5 +18,5 @@ TODO: add description
 
 <h2>Send Test Request</h2>
 <pre>
-curl {{info.base_url}}/agent/metrics -d '{"token":"{{app.token}}","metrics":[{"name":"request","value":0.15,"kind":"request"}]}'
+curl {{base_url}}/agent/metrics -d '{"token":"{{token}}","metrics":[{"name":"request","value":0.15,"kind":"request"}]}'
 </pre>
