+++ ui/app/templates/snow/apps/show.hbs
@@ -1,6 +1,6 @@
-{{#if app.requests.total.count}}
+{{#if requests.total.count}}
   <div class="page-header">
-    <h1>{{app.name}}</h1>
+    <h1>{{name}}</h1>
   </div>
 
   <div role="tabpanel">
@@ -23,7 +23,7 @@
 
 {{else}}
   <div class="page-header">
-    <h1>{{app.name}} - Installation</h1>
+    <h1>{{name}} - Installation</h1>
   </div>
   {{partial "snow/apps/install"}}
 {{/if}}
